feat(server): return epg list response with pagination info
Return a list response for epg events with information about the total
results and the offset. Also add the picon id of the channel to the epg
event.

+++ server/tvheadend/types.go
@@ -2,8 +2,8 @@ package tvheadend
 
 type (
 	GridResponse[T any] struct {
-		Entries []T `json:"entries"`
-		Total   int `json:"totalCount"`
+		Entries []T   `json:"entries"`
+		Total   int64 `json:"totalCount"`
 	}
 
 	ListResponse[T any] struct {
